requirements.txt
appdirs==1.4.2
asgi-redis==1.0.0
asgiref==1.0.0
asterisk-ami==0.1.0
attrs==16.3.0
autobahn==0.17.2
Automat==0.5.0
Babel==2.3.4
BabelDjango==0.2.2
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.5.1
blivet==0.61.15.59
Brlapi==0.6.0
channels==1.0.3
chardet==2.2.1
clufter==0.59.5
colorama==0.2.5
configobj==4.7.2
configshell-fb==1.1.18
constantly==15.1.0
coverage==3.6b3
cssselect==1.0.0
cssutils==1.0.1
cupshelpers==1.0
daphne==1.0.3
decorator==3.4.0
di==0.3
diff-match-patch==20121119
Django==1.10.6
django-absolute==0.3
django-appconf==1.0.2
django-autocomplete-light==3.2.1
django-bootstrap3==7.1.0
django-bootstrap3-datepicker==0.4.1
django-bootstrap3-datetimepicker-2==2.4.2
django-braces==1.9.0
django-ckeditor==5.2.1
django-common-helpers==0.9.0
django-countries==4.0
django-crispy-forms==1.6.0
django-cron==0.5.0
django-datatables-view==1.13.1
django-datetime-widget==0.9.3
django-discover-runner==1.0
django-easy-select2==1.3.3
django-emailit==0.2.1
django-filter==1.0.1
django-fm==0.2.5
django-gm2m==0.4.8
django-highcharts==0.1.7
django-import-export==0.4.5
django-multiselectfield==0.1.5
django-phonenumber-field==1.1.0
django-report-utils==0.3.15
django-role-permissions==1.2.3
Django-Select2==5.8.10
django-selectable==0.9.0
django-smart-selects==1.3.2
django-tables2==1.4.2
django-versatileimagefield==1.6
djangorestframework==3.5.4
docutils==0.12
ethtool==0.8
feedparser==5.1.3
firstboot==19.5
fros==1.0
gdata==2.0.18
gevent==1.0.2
greenlet==0.4.7
html5lib==0.999
incremental==16.10.1
iniparse==0.4
initial-setup==0.3.9.36
iotop==0.6
ipaddress==1.0.17
IPy==0.75
javapackages==1.0.0
jcconv==0.2.3
Jinja2==2.7.3
kitchen==1.1.1
kmod==0.1
langtable==0.0.31
lvm===2.02.166-2-RHEL7.-2016-11-16-
lxml==3.6.4
Mako==1.0.1
Markdown==2.6.8
MarkupSafe==0.23
maxminddb==1.2.1
maxminddb-geolite2==2015.708
msgpack-python==0.4.8
MySQL-python==1.2.5
ntplib==0.3.2
olefile==0.44
packaging==16.8
pam==0.1.4
pcp==1.0
pcs==0.9.152
perf==0.1
pexpect==2.3
phonenumberslite==7.7.3
Pillow==2.7.0
policycoreutils-default-encoding==0.1
premailer==3.0.1
prices==0.5.8
psycopg2==2.6.2
py-Asterisk==0.5.10
pycups==1.9.63
pycurl==7.19.0
pygobject==3.14.0
pygpgme==0.3
pyinotify==0.9.4
pykickstart==1.99.66.10
pyliblzma==0.5.3
PyMySQL==0.7.9
pyparsing==2.2.0
pyparted==3.9
pysmbc==1.0.13
python-augeas==0.5.0
Python-Chart==1.39
python-dmidecode==3.10.13
python-geoip==1.2
python-geoip-geolite2==2015.303
python-meh==0.25.2
python-memcached==1.48
python-nss==0.16.0
pytz==2014.10
pyudev==0.15
pyxattr==0.5.1
PyYAML==3.11
redis==2.10.5
requests==2.11.1
rtslib-fb==2.1.57
seobject==0.1
sepolicy==1.1
setproctitle==1.1.6
setroubleshoot==1.1
six==1.10.0
slip==0.4.0
slip.dbus==0.4.0
suds==0.4.1
tablib==0.11.2
targetcli-fb===2.1.fb41
targetd==0.7
Twisted==17.1.0
txaio==2.6.1
urlgrabber==3.10
urllib3==1.10.2
urwid==1.1.1
virtualenv==15.1.0
Werkzeug==0.9.6
yum-langpacks==0.4.2
yum-metadata-parser==1.1.4
zope.interface==4.3.3
